requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-rest-framework==0.1.0
djangorestframework-simplejwt==5.3.1
ruff==0.6.4
ruff-lsp==0.0.56
pillow~=10.4.0
pyotp==2.9.0
python-dotenv==1.0.1
django-cors-headers==4.4.0
social-auth-app-django==5.4.2
smsaero-api-async==3.0.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.45
unidecode==1.3.8
django-query-counter==0.4.0
user-agents==2.2.0
drf-spectacular==0.27.2
django-debug-toolbar==4.4.6
django-stubs==5.1.0
djangorestframework-stubs==3.15.1
